+++ src/ExceptionInterface.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ *
+ * This file is part of Aura for PHP.
+ *
+ * @license http://opensource.org/licenses/mit-license.php MIT
+ *
+ */
+namespace Aura\SqlQuery;
+
+/**
+ *
+ * Marker interface for all package-level exceptions; catch this to catch
+ * any exception thrown by this package, regardless of its SPL base class.
+ *
+ * @package Aura.SqlQuery
+ *
+ */
+interface ExceptionInterface extends \Throwable
+{
+}